requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.7.4.post0
    # via slackclient
amqp==5.0.9
    # via kombu
asgiref==3.3.4
    # via django
async-timeout==3.0.1
    # via aiohttp
attrs==20.3.0
    # via aiohttp
beautifulsoup4==4.9.3
    # via bs4
billiard==3.6.4.0
    # via celery
bs4==0.0.1
    # via -r requirements.in
celery[redis]==5.2.2
    # via -r requirements.in
certifi==2020.12.5
    # via
    #   requests
    #   sentry-sdk
chardet==4.0.0
    # via
    #   aiohttp
    #   requests
click==8.0.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   flask
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
dj-database-url==0.5.0
    # via -r requirements.in
django==3.2.15
    # via
    #   -r requirements.in
    #   django-staff-sso-client
    #   djangorestframework
django-environ==0.4.5
    # via -r requirements.in
django-ipware==3.0.2
    # via django-log-formatter-ecs
django-log-formatter-ecs==0.0.5
    # via -r requirements.in
django-staff-sso-client==3.1.0
    # via -r requirements.in
djangorestframework==3.12.4
    # via -r requirements.in
emoji==1.2.0
    # via -r requirements.in
flask==1.1.2
    # via slackeventsapi
idna==2.10
    # via
    #   requests
    #   yarl
itsdangerous==1.1.0
    # via
    #   flask
    #   slackeventsapi
jinja2==2.11.3
    # via
    #   flask
    #   slackeventsapi
kombu==5.2.3
    # via celery
kubi-ecs-logger==0.1.0
    # via django-log-formatter-ecs
markdown==3.3.4
    # via -r requirements.in
markupsafe==1.1.1
    # via
    #   jinja2
    #   slackeventsapi
marshmallow==3.6.1
    # via kubi-ecs-logger
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
oauthlib==3.1.1
    # via requests-oauthlib
pdpyras==4.1.4
    # via -r requirements.in
prompt-toolkit==3.0.18
    # via click-repl
psycopg2-binary==2.8.6
    # via -r requirements.in
pyee==7.0.4
    # via slackeventsapi
python-dateutil==2.8.1
    # via
    #   -r requirements.in
python-decouple==3.4
    # via -r requirements.in
pytz==2021.1
    # via
    #   celery
    #   django
redis==3.5.3
    # via celery
requests==2.25.1
    # via
    #   -r requirements.in
    #   pdpyras
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via django-staff-sso-client
sentry-sdk==1.0.0
    # via -r requirements.in
six==1.15.0
    # via
    #   click-repl
    #   python-dateutil
slackclient==2.9.3
    # via -r requirements.in
slackeventsapi==2.2.1
    # via -r requirements.in
soupsieve==2.2.1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
typing-extensions==3.7.4.3
    # via aiohttp
urllib3==1.26.5
    # via
    #   -r requirements.in
    #   pdpyras
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
waitress==2.1.2
    # via -r requirements.in
wcwidth==0.2.5
    # via prompt-toolkit
werkzeug==1.0.1
    # via flask
whitenoise==5.2.0
    # via -r requirements.in
yarl==1.6.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
